Remove Offset constructor
The `Offset` struct has multiple `usize` fields.  Having a constructor
makes the code harder to read because the constructor function call is
ambiguous in regards to parameter order.  Directly initializing the
structure is cleaner since we can label the fields.

Remove constructor, initialize struct directly, and while we are at it
re-order the struct fields to match the more intuitive ordering that was
imposed by the old constructor.